@@ -3,14 +3,25 @@ Path planning algorithms for IR-SIM simulation.
 
 This package contains:
 - a_star: A* path planning algorithm
+- jps: Jump Point Search (JPS) grid planning
 - rrt: Rapidly-exploring Random Tree algorithm
 - rrt_star: RRT* optimized path planning
+- informed_rrt_star: Informed RRT* optimal path planning
 - probabilistic_road_map: PRM path planning algorithm
 """
 
 from irsim.lib.path_planners.a_star import AStarPlanner
+from irsim.lib.path_planners.informed_rrt_star import InformedRRTStar
+from irsim.lib.path_planners.jps import JPSPlanner
 from irsim.lib.path_planners.probabilistic_road_map import PRMPlanner
 from irsim.lib.path_planners.rrt import RRT
 from irsim.lib.path_planners.rrt_star import RRTStar
 
-__all__ = ["RRT", "AStarPlanner", "PRMPlanner", "RRTStar"]
+__all__ = [
+    "RRT",
+    "AStarPlanner",
+    "InformedRRTStar",
+    "JPSPlanner",
+    "PRMPlanner",
+    "RRTStar",
+]
